lightnvm: define chunk states
Define chunk states as given in the 2.0 spec. Also, add an extra chunk
state that signals that the chunk is in use by the host. This allows for the
chunk metadata to be "owned" by a target when active, thus completing the chunk
state machine from the host perspective and facilitating sanity checks. This
state is transparent for the device.

Signed-off-by: Javier González <javier@cnexlabs.com>